% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBtypeR.R
\name{filter_tbtype}
\alias{filter_tbtype}
\title{Filter MTB phylotype mixture results based on provided criteria}
\usage{
filter_tbtype(
  tbtype_results,
  max_phylotypes = Inf,
  min_mix_prop = 0.01,
  max_p_val_perm = 0.01,
  max_p_val_wsrst = 0.01,
  min_abs_diff = 0
)
}
\arguments{
\item{tbtype_results}{results tibble obtained from the \code{tbtype} function.}

\item{max_phylotypes}{Maximum number of phylotypes in a mixture to retain. Default is \code{Inf}.}

\item{min_mix_prop}{Minimum proportion of a mixture to retain. Default is 0.01.}

\item{max_p_val_perm}{Maximum p-value for permutation test to retain a phylotype. Default is 0.01.}

\item{max_p_val_wsrst}{Maximum p-value for Wilcoxon signed-rank test to retain a phylotype. Default is 0.01.}

\item{min_abs_diff}{Minimum absolute difference in model BAF to return. Default is 0.}
}
\value{
A filtered tibble of MTB phylotype mixtures.
}
\description{
Filter MTB phylotype mixture results based on provided criteria
}
\examples{
\dontrun{
# Example usage:
tbtype_results <- tbtype(vcf = "example.vcf", max_phylotypes = 3L, min_mix_prop = 0.01)
filtered_results <- filter_tbtype(tbtype_results,
                                  max_phylotypes = 3,
                                  min_mix_prop = 0.02)
}
}
