% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{h37rv_core_genome}
\alias{h37rv_core_genome}
\title{MTBC core genomic regions (H37RV coordinates)}
\format{
GRanges object with 1155 ranges and 0 metadata columns
}
\source{
Pan and Core Genome Analysis of 183 Mycobacterium tuberculosis Strains Revealed a High Inter-Species Diversity among the Human Adapted Strains \url{https://doi.org/10.3390/antibiotics10050500}
}
\usage{
h37rv_core_genome
}
\description{
MTBC core genomic regions (H37RV coordinates)
}
\keyword{datasets}
