% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mtb_core_genome}
\alias{mtb_core_genome}
\title{Coordinates of core genome coding sequences in H37Rv}
\format{
A data frame with 1,155 rows and 3 columns:
\describe{
  \item{gene}{H37Rv gene name}
  \item{pos}{start position of gene}
  \item{sequence}{sequence of gene in H37Rv}
}
}
\source{
Pan and Core Genome Analysis of 183 Mycobacterium tuberculosis Strains Revealed a High Inter-Species Diversity among the Human Adapted Strains,
Zakham et al., 2021 \url{https://doi.org/10.3390/antibiotics10050500}
}
\usage{
mtb_core_genome
}
\description{
Coordinates of core genome coding sequences in H37Rv
}
\keyword{datasets}
