% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tbt_panel}
\alias{tbt_panel}
\title{TBtypeR SNP phylotype panel.}
\format{
A data frame with 10,904 rows and 8 columns:
\describe{
  \item{chrom}{H37Rv chromosome}
  \item{pos}{H37Rv genomic position}
  \item{ref}{H37Rv reference base}
  \item{alt}{H37Rv alternate base}
  \item{genotype}{The genoptype associated with the phylotype, 0 for referene, 1 for alternate}
  \item{phylotype}{Name of MTBC phylotype}
  \item{parent_phylotype}{Name of parent MTBC phylotype}
  \item{referece}{Paper referencing the SNP marker}
}
}
\source{
Napier et al. 2020: \url{https://doi.org/10.1186/s13073-020-00817-3}

Thawornwattana et al. 2021: \url{https://doi.org/10.1099/mgen.0.000697}

Coscolla et al. 2021: \url{https://doi.org/10.1099/mgen.0.000477}

Zwyer et al. 2021: \url{https://doi.org/10.12688/openreseurope.14029.2}

Shuaib et al. 2022: \url{https://doi.org/10.3390/genes13060990}
}
\usage{
tbt_panel
}
\description{
TBtypeR SNP phylotype panel.
}
\keyword{datasets}
